refactor(p2p): move main loop and clients to dedicated modules
