+++ crates/p2p/src/client.rs
@@ -1,0 +1,2 @@
+pub mod peer_agnostic;
+pub mod peer_aware;
